Add CRC32 check.

+++ src/app/message.cpp
@@ -19,7 +19,7 @@ void Message::Set(Connection* conn, const char header[8], std::string&& data_) {
   conn_ = conn;
   data_len = BytesToUint16(kLittleEndian, &header[0]);
   code = BytesToUint16(kLittleEndian, &header[2]);
-  crc32 = static_cast<int32_t>(BytesToUint32(kLittleEndian, &header[4]));
+  crc32 = BytesToUint32(kLittleEndian, &header[4]);
   data = std::move(data_);
 }
